lab1_2/backend/requirements.txt
asgiref==3.8.1
Django==5.2
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
Faker==37.1.0
gunicorn==23.0.0
Markdown==3.8
packaging==25.0
psycopg==3.2.6
psycopg-binary==3.2.6
psycopg-pool==3.2.6
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2

